+++ core/core-auth/build.gradle
@@ -1,5 +1,5 @@
 dependencies {
-    compileOnly("org.springframework:spring-context")
-    implementation 'com.auth0:java-jwt:4.4.0'
+    compileOnly 'org.springframework.boot:spring-boot-starter-web'
 
+    implementation 'com.auth0:java-jwt:4.4.0'
 }
