+++ core/core-auth/src/main/java/com/barlow/core/support/crypto/AlgorithmConfig.java
@@ -1,4 +1,4 @@
-package com.barlow.core.auth.support.crypto;
+package com.barlow.core.support.crypto;
 
 import org.springframework.beans.factory.BeanCreationException;
 import org.springframework.beans.factory.annotation.Value;
