+++ core/core-auth/src/main/java/com/barlow/core/support/crypto/KeyAlgorithm.java
@@ -1,4 +1,4 @@
-package com.barlow.core.auth.support.crypto;
+package com.barlow.core.support.crypto;
 
 import com.auth0.jwt.algorithms.Algorithm;
