+++ core/core-auth/src/main/java/com/barlow/core/support/jwt/JwtConfig.java
@@ -1,4 +1,4 @@
-package com.barlow.core.auth.support.jwt;
+package com.barlow.core.support.jwt;
 
 
 import java.time.Duration;
