+++ core/core-auth/src/test/java/com/barlow/core/auth/authentication/token/AccessTokenAuthenticatorTest.java
@@ -5,8 +5,8 @@ import com.barlow.core.auth.authentication.core.AuthenticationException;
 import com.barlow.core.auth.authentication.core.AuthenticationExceptionType;
 import com.barlow.core.auth.authentication.core.MemberPrincipal;
 import com.barlow.core.auth.config.TestKeyConfig;
-import com.barlow.core.auth.support.crypto.PrivateKeyAlgorithm;
-import com.barlow.core.auth.support.jwt.TokenConfig;
+import com.barlow.core.support.crypto.PrivateKeyAlgorithm;
+import com.barlow.core.support.jwt.TokenConfig;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
