+++ core/core-auth/src/test/java/com/barlow/core/auth/authentication/token/AccessTokenProviderTest.java
@@ -5,8 +5,8 @@ import com.auth0.jwt.JWTVerifier;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import com.barlow.core.auth.authentication.core.MemberPrincipal;
 import com.barlow.core.auth.config.TestKeyConfig;
-import com.barlow.core.auth.support.crypto.PublicKeyAlgorithm;
-import com.barlow.core.auth.support.jwt.TokenConfig;
+import com.barlow.core.support.crypto.PublicKeyAlgorithm;
+import com.barlow.core.support.jwt.TokenConfig;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
