+++ core/core-auth/src/test/java/com/barlow/core/auth/support/crypto/AlgorithmConfigTest.java
@@ -2,6 +2,10 @@ package com.barlow.core.auth.support.crypto;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
+import com.barlow.core.support.crypto.AlgorithmConfig;
+import com.barlow.core.support.crypto.PrivateKeyAlgorithm;
+import com.barlow.core.support.crypto.PublicKeyAlgorithm;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
